Internals

Info

The following interfaces are used internally and as such aren't exported by ChemEquations. However, you can still access them with using ChemEquations: interfacename.

ChemEquations.CHARGEREGEX — Constant

CHARGEREGEX

Regex to match {...} charge string.

ChemEquations.EQUALCHARS — Constant

EQUALCHARS

Characters used to split a chemical equation into two sides. Source: https://github.com/SciML/Catalyst.jl/blob/master/src/reaction_network.jl#L56

ChemEquations.PLUSREGEX — Constant

PLUSREGEX

Regex to split a chemical equation into compounds.

ChemEquations.CompoundTuple — Type

CompoundTuple

Type stored in ChemEquation.tuples.

ChemEquations.ElementTuple — Type

ElementTuple

Type stored in Compound.tuples.

ChemEquations.charge — Method

charge(str::AbstractString, T::Type) -> Any

Extracts charge from compound's string into a number of specified type.

ChemEquations.compoundtuples — Method

compoundtuples(str::AbstractString, T::Type) -> Array{_A,1} where _A

Extracts compound tuples from equation's string.

ChemEquations.elementtuples — Method

elementtuples(str::AbstractString, T::Type) -> Array{_A,1} where _A

Extracts element tuples from compound's string.