Public Interface

ChemEquations.ChemEquations — Module

ChemEquations

Write and balance chemical equations elegantly and efficiently.

Types

ChemEquations.ChemEquation — Type

struct ChemEquation{T<:Number}

tuples::Array{Tuple{Compound{T},T},1} where T<:Number

Stores chemical equation's compounds and their coefficients in a structured way.

ChemEquations.ChemEquation — Method

ChemEquation(str::AbstractString) -> ChemEquation{Int64}

Constructs a chemical equation from the given string.

Parsing is insensitive to whitespace. Any character in EQUALCHARS separates the equation into two sides, while + separates the equation into compounds.

Examples

julia> ChemEquation("N2+O2⇌2NO")
ce"N2 + O2 = 2 NO"

julia> ChemEquation("CH3COOH + Na → H2 + CH3COONa")
ce"C2H4O2 + Na = H2 + C2H3O2Na"

julia> ChemEquation("⏣H + Cl2 = ⏣Cl + HCl")
ce"⏣H + Cl2 = ⏣Cl + HCl"

ChemEquations.Compound — Type

struct Compound{T<:Number}

tuples::Array{Tuple{String,T},1} where T<:Number

charge::Number

Stores chemical compound's elements and charge in a structured way.

Info

Electron is stored as "e".

ChemEquations.Compound — Method

Compound(str::AbstractString) -> Compound{Int64}

Constructs a compound from str.

An element begins with an uppercase unicode letter and ends with a lowercase unicode letter or a unicode symbol.

Info

An element can also begin with a symbol if the symbol is the first character (e.g. "⬡H").

Parsing is insensitive to whitespace and underscores (_). Special parsing is implemented for:

parens (e.g. "(CH3COO)2Mg")

compounds with "*" (e.g. "CuSO4 * 5H2O")

electrons ("e")

Charge is in the form "{±n}" or "{n±}". It is automatically deduced for electron ("e").

Examples

julia> Compound("H2O")
cc"H2O"

julia> Compound("H3O{+}")
cc"H3O{+}"

julia> Compound("(CH3COO)2Mg")
cc"C4H6O4Mg"

julia> Compound("CuSO4 * 5H2O")
cc"CuSO9H10"

julia> Compound("⬡Cl")
cc"⬡Cl"

Macros

ChemEquations.@cc_str — Macro

@cc_str

Constructs a compound with cc"str" syntax, instead of Compound(str).

Examples

julia> cc"H3O{+1}"
cc"H3O{+}"

ChemEquations.@ce_str — Macro

@ce_str

Constructs a chemical equation with ce"str" syntax, instead of ChemEquation(str).

Examples

julia> ce"H2 + O2 → H2O"
ce"H2 + O2 = H2O"

Functions

Base.:== — Method

==(equation_1::ChemEquation, equation_2::ChemEquation) -> Any

Checks whether two equations are chemically equal.

Examples

julia> ce"H2 + O2 = H2O" == ce"O2 + H2 → H2O"
true

Base.:== — Method

==(compound_1::Compound, compound_2::Compound) -> Any

Checks whether two compounds are chemically equal.

Examples

julia> cc"MgOHOH" == cc"Mg(OH)2"
true

Base.show — Method

show(io::IO, equation::ChemEquation)

Displays the chemical equation using Base.string(::Compound).

Base.show — Method

show(io::IO, compound::Compound)

Displays the compound using Base.string(::Compound).

Base.string — Method

string(equation::ChemEquation) -> String

Creates a string to represent the chemical equation.

All compounds are displayed with Base.string(::Compound), in the order in which they were originally given, with coefficients equal to 1 not displayed. '=' and '+' are used as separators, with spaces inserted for easier reading.

Examples

julia> string(ce"Cr2O7{2-} + H{+} + {-} = Cr{3+} + H2O")
"Cr2O7{-2} + H{+} + e = Cr{+3} + H2O"

Base.string — Method

string(compound::Compound) -> Union{Missing, Regex, String}

Creates a string to represent the compound.

All elements are displayed only once (e.g. "H2O" and not "HOH"), in the order in which they were originally given (e.g. "MgO2H2" from cc"Mg(OH)2"), with coefficients equal to 1 not displayed (e.g. "O" and not "O1").

Examples

julia> string(cc"CuSO4 * 5 H2O")
"CuSO9H10"

ChemEquations.balance — Method

balance(mat::AbstractArray{T,2} where T) -> Any

Balances an equation matrix using the nullspace method. Returns an array in which each column represents a solution.

References

Thorne (2009)

ChemEquations.balance — Method

balance(equation::ChemEquation; fractions) -> ChemEquation

Balances the coefficients of a chemical equation. The minimal integer solution is always displayed.

If the equation cannot be balanced, an error is thrown.

Info

The original equation is not modified.

Examples

julia> balance(ce"Fe + Cl2 = FeCl3")
ce"2 Fe + 3 Cl2 = 2 FeCl3"

julia> balance(ce"Cr2O7{-2} + H{+} + {-} = Cr{+3} + H2O")
ce"Cr2O7{-2} + 14 H{+} + 6 e = 2 Cr{+3} + 7 H2O"

ChemEquations.compounds — Method

compounds(equation::ChemEquation) -> Array{_A,1} where _A

Returns chemical equation's compounds in a list.

ChemEquations.elements — Method

elements(equation::ChemEquation) -> Array{String,1}

Returns chemical equation's unique elements.

Examples

julia> elements(ce"2 H2 + O2 → 2 H2O")
2-element Array{String,1}:
 "H"
 "O"

ChemEquations.elements — Method

elements(compound::Compound) -> Any

Returns compound's elements as strings.

Examples

julia> elements(cc"CH3COOH")
3-element Array{String,1}:
 "C"
 "H"
 "O"

ChemEquations.equationmatrix — Method

equationmatrix(equation::ChemEquation{T}) -> Array

Creates an equation matrix in which rows correspond to atoms and columns correspond to compounds.

Examples

julia> equationmatrix(ce"H2 + Cl2 → HCl")
2×3 Array{Int64,2}:
 2  0  1
 0  2  1

ChemEquations.hascharge — Method

hascharge(equation::ChemEquation) -> Bool

True if chemical equation has at least one compound with nonzero charge.

ChemEquations.hascharge — Method

hascharge(compound::Compound) -> Union{Missing, Bool, Base.var"#64#65"{_A} where _A}

True if the compound's charge is nonzero.